+++ CHANGELOG.md
@@ -8,12 +8,14 @@ and this project adheres to
 ## [Unreleased]
 ### Added
 - `watchMode.includeSourceDirectories` option to replace `watchSourceDirectories`
+- `watchMode.alwaysRun` option to replace `always`
 - `watchMode.skipInitialRun` option to skip task during inital run while in watch mode
 - `watchMode.changedFilesOnly` option only include changed files in task stream while in watch mode
 - Additional keywords to `package.json`
 
 ### Changed
 - Deprecate `watchSourceDirectories` option in favor of `watchMode.includeSourceDirectories`
+- Deprecate `always` option in favor of `watchMode.alwaysRun`
 
 ## [0.1.1] - 2019-12-31
 ### Added